3.2 updates

+++ massiveAdmin/class/massiveAdmin.class.php
@@ -461,26 +461,6 @@ class MassiveAdminClass
     {
 
 
-        $url = $_POST['url'];
-
-        file_put_contents(GSPLUGINPATH . "Tmpfile.zip", fopen("$url", 'r'));
-        $path = GSPLUGINPATH . "Tmpfile.zip";
-
-        $zip = new ZipArchive;
-        if ($zip->open($path) === TRUE) {
-            if (file_exists(GSPLUGINPATH . "tmp_plugin/") == false) {
-                mkdir(GSPLUGINPATH . "tmp_plugin/", 0755);
-            };
-            $zip->extractTo(GSPLUGINPATH  . "tmp_plugin/");
-            $zip->close();
-
-            foreach (glob(GSPLUGINPATH  . "tmp_plugin/*/*") as $filename) {
-                if (file_exists(str_replace(pathinfo($filename)['dirname'], GSPLUGINPATH, $filename)) !== true) {
-                    rename($filename, str_replace(pathinfo($filename)['dirname'], GSPLUGINPATH, $filename));
-                };
-            };
-
-
 
             function delete_directory($dirname)
             {
@@ -501,6 +481,33 @@ class MassiveAdminClass
                 return true;
             };
 
+
+        $url = $_POST['url'];
+
+        file_put_contents(GSPLUGINPATH . "Tmpfile.zip", fopen("$url", 'r'));
+        $path = GSPLUGINPATH . "Tmpfile.zip";
+
+        $zip = new ZipArchive;
+        if ($zip->open($path) === TRUE) {
+            if (file_exists(GSPLUGINPATH . "tmp_plugin/") == false) {
+                mkdir(GSPLUGINPATH . "tmp_plugin/", 0755);
+            };
+            $zip->extractTo(GSPLUGINPATH  . "tmp_plugin/");
+            $zip->close();
+
+            foreach (glob(GSPLUGINPATH  . "tmp_plugin/*/*") as $filename) {
+           
+           if(file_exists(str_replace(pathinfo($filename)['dirname'], GSPLUGINPATH,$filename))){
+           delete_directory(str_replace(pathinfo($filename)['dirname'], GSPLUGINPATH,$filename));
+           }
+
+
+                    rename($filename, str_replace(pathinfo($filename)['dirname'], GSPLUGINPATH, $filename));
+                 
+            };
+
+
+
             delete_directory(GSPLUGINPATH . "tmp_plugin");
 
             unlink($path);
@@ -511,6 +518,69 @@ class MassiveAdminClass
         echo ("<meta http-equiv='refresh' content='1'>");
     }
 
+
+//remover
+
+    public function unistaller(){
+        
+$delPlug = $_GET['delPlugin'];
+
+
+            function delete_directory($dirname)
+            {
+              
+              if(file_exists($dirname)){
+
+
+  if (is_dir($dirname))
+                    $dir_handle = opendir($dirname);
+                if (!$dir_handle)
+                    return false;
+                while ($file = readdir($dir_handle)) {
+                    if ($file != "." && $file != "..") {
+                        if (!is_dir($dirname . "/" . $file))
+                            unlink($dirname . "/" . $file);
+                        else
+                            delete_directory($dirname . '/' . $file);
+                    }
+                }
+                closedir($dir_handle);
+                rmdir($dirname);
+                return true;
+
+              }
+
+            };
+
+if(GSPLUGINPATH.$delPlug){
+delete_directory(GSPLUGINPATH.$delPlug);    
+};
+
+if(file_exists(GSPLUGINPATH.$delPlug.'.php')){
+    unlink(GSPLUGINPATH.$delPlug.'.php');
+}
+ global $GSADMIN;
+global $SITEURL;
+$url =  $SITEURL.$GSADMIN.'/load.php?id=massiveAdmin&unistaller';
+
+      echo '<div class="success" style="
+      position:absolute;top:0;left:0;
+    width: 100%;
+    background: green;
+    padding: 10px;
+    box-sizing: border-box;
+    color: #fff;
+    margin-bottom: 20px;
+      ">Removed!</div>';
+        echo ("<script>
+
+setTimeout(()=>{
+
+window.location.href = '".$url."';
+    },1000);
+
+            </script>");
+    }
 
     //snippet save
